servicecloud/incident_classifier/oci/classifier/requirements.txt
oracle-ads==2.6.3
pandas==1.3.5
beautifulsoup4==4.11.1
python-dateutil==2.8.2
pycm==3.6
html5lib==1.1
cloudpickle==2.1.0
scikit-learn==1.0.2
numpy==1.22.2
joblib==1.1.0
oci==2.79.0
requests==2.28.1
scipy==1.9.0
ipython==8.4.0
schedule==1.1.0
pytest==7.1.2
six==1.16.0
PyYAML==5.4.1
lxml==4.7.1
jwt==1.3.1
PyJWT==2.4.0
spacy==3.4.1
